+++ services/api/src/offers/offers.module.ts
@@ -7,6 +7,7 @@ import { OffersController } from './offers.controller';
 import { OffersService } from './offers.service';
 import { OffersRepository } from './offers.repository';
 import { CommissionService } from './commission/commission.service';
+import { CentrifugoClient } from './delivery/centrifugo.client';
 import { DeliverySchedulerService } from './delivery/delivery-scheduler.service';
 import { OfferNotificationService } from './notification/offer-notification.service';
 import { OfferEventEmitterService } from './events/offer-event-emitter.service';
@@ -46,6 +47,7 @@ import { OfferDelivery } from './entities/offer-delivery.entity';
     OffersService,
     OffersRepository,
     CommissionService,
+    CentrifugoClient,
     DeliverySchedulerService,
     OfferNotificationService,
     OfferEventEmitterService,